requirements.txt
Orange3
simple-salesforce
requests
